cerbero: Return libname:filename dict from library search function
While generating a .lib library file on Windows, we guess the library name from
the DLL filename, but we often get it wrong since the code that guesses is naive
and assumes that the DLL is either of the form libfoo.dll or libfoo-N.dll.
However, nettle.recipe installs libfoo-N-M.dll and we get the library name
wrong.

We don't need to guess the name if we remember and pass the library name
correctly, so let's just do that.